change to using FIFO stock system

+++ database/migrations/2025_02_09_130449_create_pembelians_table.php
@@ -16,7 +16,7 @@ class CreatePembeliansTable extends Migration
         Schema::create('pembelians', function (Blueprint $table) {
             $table->id();
             $table->date('tanggal_beli');
-            $table->integer('total');
+            $table->decimal('total', 15, 2);
             $table->foreignId('suppliers_id')->constrained('suppliers');
             $table->foreignId('users_id')->constrained('users');
             $table->timestamps();
